etl/pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = ["mishamsk <mishamsk@gmail.com>"]
packages = [{ include = "loader" }]

[tool.poetry.dependencies]
python = "^3.10"
zeep = "^4.1.0"
click = "^8.1.3"
dbt-postgres = "^1.3.0"
drebedengi-python-api = "^0.1.0"
psycopg2 = "^2.9.4"
SQLAlchemy = { extras = ["mypy"], version = "^1.4.41" }
python-dotenv = "^0.20.0"
alembic = "^1.8.1"
PyYAML = "^6.0"
dataclasses-json = "^0.5.7"
nerddiary = "^0.3.2"
requests = "^2.28.1"

[tool.poetry.scripts]
loader = 'loader.cli:run_cli'

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"
